+++ src/test/java/backend/wal/auth/application/service/IssueTokenServiceTest.java
@@ -2,7 +2,7 @@ package backend.wal.auth.application.service;
 
 import backend.wal.advice.exception.NotFoundException;
 import backend.wal.auth.application.port.out.CreateRefreshTokenResponseDto;
-import backend.wal.auth.application.port.out.TokenResponseDto;
+import backend.wal.auth.application.port.in.TokenResponseDto;
 import backend.wal.auth.application.port.out.JwtManagerPort;
 import backend.wal.auth.domain.RefreshToken;
 import backend.wal.auth.domain.repository.RefreshTokenRepository;
